seperate title and branding into new component
